abc297

+++ simu/grid_abc88.cpp
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
-#define rep(i,e) for(int i=0;i<(e);i++)
+#define rep(i,e) for(int i=0;i<(int)(e);i++)
 using ll=long long int;
 using P=pair<int,int>;
